requirements.txt
streamlit==1.29.0
supabase==2.3.0
pandas==2.1.4
python-dotenv==1.0.0
plotly==5.17.0
datetime